Move getStats function to task service
